Changes to trajectory generation, evaluating a trajectory, and launch files
Trajectory Generation

The changes for degree of freedoms in a trajectory are no longer generated simultaneously. The trajectories are generated in three steps: 1) orientation change to rotate towards the goal, 2) position change to drive straight towards the goal, 3) orientation change to rotate to the goal orientation.
The Segment class now has members: T_rotate_pre_ - the time it takes to rotate towards the goal, T_loc_ - the time it takes to drive straight towards the goal, T_rotate_post_ - the time it takes to rotate to the goal orientation, and T_min_ - the time it takes to complete the segment, which is the sum of the previous three members. The T_ vector has been removed.

Evaluation

Collision detection has been changed from line intersection to circle intersection.

Launch Files

The launch files have been changed to start all necessary nodes on two robots. Resetting the router changed the robots' ip addresses on the wlan so those had to be adjusted.

Robot package

Minor changes. The next commit should have the robots moving along a new trajectory.